Moving to a disclaimer component.
Mostly just messing around with the UI a bit
