+++ src/components/Charts/ModelChart.js
@@ -2,8 +2,6 @@ import React, { useMemo } from 'react';
 import { dateFormat } from 'highcharts';
 import moment from 'moment';
 import { INTERVENTIONS } from 'enums/interventions';
-import LightTooltip from 'components/LightTooltip/LightTooltip';
-import ClaimStateBlock from 'components/ClaimStateBlock/ClaimStateBlock';
 import Chart from './Chart';
 import { isEmpty } from 'lodash';
 import { COLOR_MAP } from 'enums/interventions';
@@ -12,10 +10,6 @@ import ReactDOMServer from 'react-dom/server';
 import {
   ChartContainer,
   Wrapper,
-  DisclaimerWrapper,
-  Disclaimer,
-  DisclaimerHeader,
-  DisclaimerBody,
   CondensedLegendStyled,
   CondensedLegendItemStyled,
 } from './ModelChart.style';
@@ -41,10 +35,8 @@ const ModelChart = ({
   condensed,
   projections,
   currentIntervention,
-  lastUpdatedDate,
   forCompareModels, // true when used by CompareInterventions.js component.
-  stateId,
-  selectedCounty,
+  children,
 }) => {
   // We use the inferred projection if supported, otherwise the worst case for the currently active intervention
   let projection = projections.primary;
@@ -325,36 +317,7 @@ const ModelChart = ({
     <ChartContainer>
       <Wrapper projections={projections} isInferred={projection.isInferred}>
         <Chart options={options} />
-        <DisclaimerWrapper>
-          <Disclaimer>
-            <DisclaimerHeader>
-              <LightTooltip
-                title="Currently we aggregate data over 4 day intervals to smooth out inconsistencies in the source data. We’re working on improving this now."
-                placement="bottom"
-              >
-                <span>
-                  Last updated{' '}
-                  {lastUpdatedDate && lastUpdatedDate.toLocaleDateString()}
-                </span>
-              </LightTooltip>
-            </DisclaimerHeader>
-            <DisclaimerBody>
-              This model updates every 3 days and is intended to help make fast
-              decisions, not predict the future.{' '}
-              <a
-                href="https://data.covidactnow.org/Covid_Act_Now_Model_References_and_Assumptions.pdf"
-                target="_blank"
-                rel="noopener noreferrer"
-              >
-                Learn more about our projection and its limitations
-              </a>
-              .
-            </DisclaimerBody>
-          </Disclaimer>
-          <Disclaimer>
-            <ClaimStateBlock stateId={stateId} county={selectedCounty} />
-          </Disclaimer>
-        </DisclaimerWrapper>
+        {children}
       </Wrapper>
     </ChartContainer>
   );
